Path to v0.1
JS, Generator, Tool: Allows for arbitrary sizes of any kind. In the config JSON, in the "ImageSizes" field, give it a key/value that has the size name as the key, and an ImageScale object as the value. This will translate to the tool creating a directory in the ImageRootDirectory of every folder with the given size.

JS: You can now set where thumbnails are displayed from. It must be a valid folder in the ImageRootDirectory, otherwise it will 404.

Generator: Implements more Go testing

JS, Generator: WebConfig implements WebImageSize for the arbitrary sizing. Each size consists of SizeName, Directory, and LocalBool - SizeName and Directory are equal, while LocalBool tells the script whether to fetch the photo from either the local server, or from a remote server (as defined in PhotoURLBase)

Tool: Implements Go testing

Tool: GeneratorOptions now has three new boolean options: sort, imagegen, and gensizes. sort sorts the images in a folder, imagegen allows images to be generated (currently only used in GenerateFolder), and gensizes generates all possible sizes of image as defined in ImageSizes.