Path to v0.1
Generator, Tool: Implements theme configuration on generation. Not fully
implemented yet, however, and is subject to change - but themes now
require a theme.json in order to be copied over properly into the
webroot.

Generator, Tool: Implements metadata. Right now, this is subject to only
name/description. EXIF data is not copied over, and must be handled by
the front-end - this will not be implemented in fotoDen.js, and will
most likely require an external library.

Theme (default): Minor adjustments