Path to v0.1
JS, Tool: You no longer have to 'initialize' fotoDen with the fotoDen
tool. BaseURL is now grabbed from the HTML - this requires the fotoDen
script to have the ID 'fd-script', and data attribute fd-baseURL equal
to the base URL of the gallery.

JS, Generator: WorkingDirectory is now deprecated. It has been replaced
with an in-script detection function.

JS, Theme: Small bugfixes