Path to v0.1
General: A makefile has now been added, to allow for automated building
of fotoDen. The makefile includes three modes, and an 'all' target:
- minjs, which creates a minified version of fotoDen
- theme, which packages a theme and minifies all related js files
- tool, which makes the fotoDen tool with the fotoDen.js checksums
included in the file

All Go Source: Linting

JS files: Adds the @license tag so minifiers actually keep the license

Tool: Commands are now organized in a more 'ergonomic' way - flags are
now isolated to what applies to each command, and subcommands are now
implemented (e.g., init site is now a command on its own)

Tool: Initializing a fotoDen JS is now a straight copy - but in
addition, it also verifies the JS file to the version that the tool was
built with (e.g., fotoDen v0.1.0 will have its md5 checksums linked to
fotoDen.js v0.1.0). It also accepts a minified version's checksum, which
must also be built with fotoDen.